clean up a 64-bit warning
git-svn-id: https://svn.apache.org/repos/asf/httpd/httpd/branches/1.3.x@78864 13f79535-47bb-0310-9956-ffa450edef68

+++ src/modules/standard/mod_setenvif.c
@@ -161,7 +161,7 @@ static const char *add_setenvif(cmd_parms *cmd, void *mconfig, const char *args)
 	(sei_entry *)sconf->conditionals->elts;
     char *var;
     int i;
-    int cflags = (int)cmd->info;
+    int cflags = (int)(long)cmd->info;
     char *error;
     int beenhere = 0;
